requirements.txt
# NVIDIA Nsight Systems 输出文件解析工具依赖
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
numpy>=1.20.0

# LangChain + Vector store (FAISS) + Embeddings
langchain-community>=0.0.30
langchain-text-splitters>=0.0.1
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4


# 可选依赖
# sqlite3  # Python 内置模块
# json     # Python 内置模块
# csv      # Python 内置模块
# subprocess  # Python 内置模块


